moving the beam_to_xarray tto the custom_devices file, adding the ari_model_origins dictionary to ari_sim and updating some doc strings in transformations
